\name{synlik-class}
\alias{synlik-class}
\title{\code{synlik-class}}
\usage{
  .synlik(...)
}
\description{
  \describe{ \item{param}{Vector of parameters used by
  object@simulator (\code{numeric}).}
  \item{simulator}{Function that simulates from the model
  (\code{function}).} \item{summaries}{Function that
  transforms simulated data into summary statistics
  (\code{function}).} \item{data}{Vector containing the
  observed data (\code{numeric}).} \item{extraArgs}{List
  containing all the extra arguments to be passed to
  object@simulator and object@summaries (\code{list}).} }
}
\author{
  Matteo Fasiolo <matteo.fasiolo@gmail.com>
}

