\name{synlik-package}
\alias{synlik-package}
\alias{synlik}
\docType{package}
\title{
Synthetic Likelihood methods for intractable likelihoods
}
\description{
Package that provides Synthetic Likelihood tools that can be used to do simulation-based inference
for models where the likelihood function is unavailable or intractable.
}
\details{
\tabular{ll}{
Package: \tab synlik\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2013-09-10\cr
License: \tab GPL (>=2)\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Matteo Fasiolo and Simon Wood

Maintainer: Matteo Fasiolo <matteo.fasiolo@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ Intractable Likelihood, Simulation-based inference. }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
